Always show PlayerNameRow as button
